pip_requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
discord.py==2.4.0
edge-tts==6.1.15
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
PyNaCl==1.5.0
pyttsx3==2.98
PyYAML==6.0.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
yarl==1.17.1
